testutil 작성 + network 나누기
