+++ cmd/network/subnet_ip_capacity.go
@@ -2,16 +2,20 @@ package network
 
 import (
 	"context"
-	"eks-checklist/cmd"
 	"log"
 	"math"
 	"net"
 
 	"github.com/aws/aws-sdk-go-v2/service/ec2"
+	"github.com/aws/aws-sdk-go-v2/service/eks/types"
 )
 
+type EksCluster struct {
+	Cluster *types.Cluster
+}
+
 // EKS 클러스터가 사용하는 서브넷의 가용가능한 IP 주소가 10% 미만인 경우 SubnetId와 사용가능한 IP 개수를 반환하는 함수
-func CheckVpcSubnetIpCapacity(eksCluster cmd.EksCluster) map[string]int {
+func CheckVpcSubnetIpCapacity(eksCluster EksCluster) map[string]int {
 	// EKS가 배포된 VPC의 ID 및 서브넷 가져오기
 	subnetIds := eksCluster.Cluster.ResourcesVpcConfig.SubnetIds
